refactor: 선택한 셀 주변 셀들을 반환하는 책임을 Position에게 전가
